requirements.txt
Pillow==10.3.0
opencv-python==4.9.0.80
numpy==1.26.4